remove posts/tags/ endpoint
